Dev: move to Trait for console
