+++ application/core/LSYii_Application.php
@@ -18,6 +18,7 @@
  */
 
 require_once(dirname(dirname(__FILE__)) . '/helpers/globals.php');
+require_once __DIR__ . '/Traits/LSApplicationTrait.php';
 
 /**
 * Implements global config
@@ -36,6 +37,8 @@ require_once(dirname(dirname(__FILE__)) . '/helpers/globals.php');
 */
 class LSYii_Application extends CWebApplication
 {
+    use LSApplicationTrait;
+
     protected $config = array();
 
     /**
@@ -713,26 +716,5 @@ class LSYii_Application extends CWebApplication
             $baseUrl = $sPublicUrl;
         }
         return $baseUrl;
-    }
-
-    /**
-     * get the current id of connected user,
-     * check if user exist before return for security
-     * @return int|null user id
-     */
-    public function getCurrentUserId()
-    {
-        if(empty(App()->session['loginID'])) {
-            // NULL for guest
-            return App()->session['loginID'];
-        }
-        if (!is_null($this->currentUserId) && $this->currentUserId == App()->session['loginID']) {
-            return $this->currentUserId;
-        }
-        $this->currentUserId = App()->session['loginID'];
-        if ($this->currentUserId && !User::model()->findByPk($this->currentUserId)) {
-            $this->currentUserId = 0;
-        }
-        return $this->currentUserId;
     }
 }
